% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_score.R
\name{calCost1}
\alias{calCost1}
\title{Calculate the cost of changing one annotation to the other}
\usage{
calCost1(
  l1,
  l2,
  m = matrix(data = c(1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0,
    0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1),
    nrow = 7)
)
}
\arguments{
\item{l1}{IU list1: a list of intonation units}

\item{l2}{IU list1: a list of intonation units}

\item{m}{similarity matrix to customize substitution cost}
}
\value{
cost
}
\description{
Calculate the cost of changing one annotation to the other
}
\examples{
m=matrix(data =
c(1,0,0,0,0,0,0,
0,1,0,0,0,0,0,
0,0,1,0,0,0,0,
0,0,0,1,0,0,0,
0,0,0,0,1,0,0,
0,0,0,0,0,1,0,
0,0,0,0,0,0,1), nrow=7)
}
