% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffCompare.R
\name{compareDiff}
\alias{compareDiff}
\title{Compare Difference of Words Between Two Annotations}
\usage{
compareDiff(data1, data2)
}
\arguments{
\item{data1}{annotation_1 from read_csv,the result after processing a .rez through the rezTrans file.}

\item{data2}{annotation_2 from read_csv, the result after processing a .rez through the rezTrans file.}
}
\value{
A dataframe of different lines. If the two files are the same, return null.
}
\description{
Compare Difference of Words Between Two Annotations
}
