% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nccu_squareno}
\alias{nccu_squareno}
\title{IU segmentations of squared-numbered texts in the NCCU Taiwan Mandarin Corpus}
\format{
## `nccu_squareno`
A list of 7 items, each of which contains two items (annotations from each transcriber).
}
\usage{
nccu_squareno
}
\description{
IU segmentations of squared-numbered texts in the NCCU Taiwan Mandarin Corpus
}
\keyword{datasets}
