% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\docType{data}
\name{nccu_t049}
\alias{nccu_t049}
\title{Two separate intonation unit segmentations from text 049 of the
NCCU corpus.}
\format{
An object of class \code{list} of length 2.
}
\usage{
nccu_t049
}
\description{
Two separate intonation unit segmentations from text 049 of the
NCCU corpus.
}
\keyword{datasets}
