% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezTrans.R
\name{rezTrans}
\alias{rezTrans}
\title{Transform .rez file to the required format}
\usage{
rezTrans(inPath)
}
\arguments{
\item{inPath}{The string of the path of the .rez file}
}
\value{
A data frame that can be used for similarity score and inter-annotated agreement
}
\description{
Transform .rez file to the required format
}
