% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sbc_iubound_prediction}
\alias{sbc_iubound_prediction}
\title{Santa Barbara Corpus prediction data}
\format{
## `who`
A list of two items:
\describe{
  \item{new}{Model predictions about IU boundaries.}
  \item{old}{Original IU boundaries.}
}

Each item is a list containing 60 dataframes. Each dataframe has the following content:
A list of two items:
\describe{
  \item{new}{The document ID.}
  \item{old}{The participant label.}
  \item{unitId / newUID}{The unit ID.}
  \item{Utterance}{The text of that unit.}
}
}
\usage{
sbc_iubound_prediction
}
\description{
A
}
\keyword{datasets}
