% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_score.R
\name{sim_Score}
\alias{sim_Score}
\title{Similarity score calculation}
\usage{
sim_Score(
  d1,
  d2,
  record = FALSE,
  m = NA,
  transCost = 0.5,
  boundaries = c(",", ".", "?", "-", "+"),
  noboundary = ";",
  trans = TRUE
)
}
\arguments{
\item{d1}{A data.frame of the first annotator's annotation. Each line represents a segment. Space is used for tokenisation, which may be spaces in the case of intonation unit segmentation, turn constructional units for turn segmentation, and so on.}

\item{d2}{A data.frame of the second annotator's annotation, similar to `d1`.}

\item{record}{Whether you want to get the step of transformation (slow process!).}

\item{m}{A similarity matrix to customize substitution cost. The size of the matrix should either be the number of boundary types in `boundaries` plus two, if `noboundary` has been set, or the number of boundary types in `boundaries` plus one, otherwise. In both cases, the final column gives deletion cost, and the final row gives insertion cost. In the first case, the second-last row and column are for unclassified boundaries.}

\item{transCost}{a transposition cost: either a single value, or a vector with the same length as the number of rows/columns in `m`.}

\item{boundaries}{A vector of boundary symbols that will exist in the data.}

\item{noboundary}{A symbol assigned for unclassified boundary types. This will be appended to lines that do not end in any symbol found in `boundaries`. Use "" if unclassified boundaries are not allowed; lines not ending with a defined boundary type will then be treated as not ending in a boundary.}

\item{trans}{If `TRUE`, the transposition operation will be performed..}
}
\value{
similarity score
}
\description{
Similarity score calculation
}
\examples{
sim_Score(nccu_t049_1, nccu_t049_2, record = T)
}
