% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezTrans.R
\name{rezTrans}
\alias{rezTrans}
\title{Inter-annotator agreement}
\usage{
rezTrans(inPath)
}
\arguments{
\item{inPath}{The string of the path of the .rez file}

\item{outPath}{The string of the path of the output csv file}
}
\value{
A data frame that can be used for similarity score and inter-annotated agreement
}
\description{
Inter-annotator agreement
}
