% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_sim_score.R
\name{sim_Score}
\alias{sim_Score}
\title{Similarity score calculation}
\usage{
sim_Score(
  d1,
  d2,
  record = FALSE,
  m = matrix(data = c(1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0,
    0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1),
    nrow = 7)
)
}
\arguments{
\item{d1}{annotation_1 from read_csv}

\item{d2}{annotation_2 from read_csv}

\item{record}{whether you want to get the step of transformation (slow process)}

\item{m}{similarity matrix to customize substitution cost}
}
\value{
similarity score
}
\description{
Similarity score calculation
}
\examples{
m=matrix(data =
c(1,0,0,0,0,0,0,
0,1,0,0,0,0,0,
0,0,1,0,0,0,0,
0,0,0,1,0,0,0,
0,0,0,0,1,0,0,
0,0,0,0,0,1,0,
0,0,0,0,0,0,1), nrow=7)
}
