% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_score.R
\name{sim_Score}
\alias{sim_Score}
\title{Similarity score calculation}
\usage{
sim_Score(
  d1,
  d2,
  record = FALSE,
  m = matrix(data = c(1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0,
    0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1),
    nrow = 7),
  boundaries = c(",", ".", "?", "-", "+"),
  noboundary = ";"
)
}
\arguments{
\item{d1}{annotation_1 from read_csv, there are three columns in the csv files (Turn Speaker Utterance), each line is an Intonation Unit, space is used for tokenization, 'punctuation' are IU boundaries and should go to the end of each IU}

\item{d2}{annotation_2 from read_csv, there are three columns in the csv files (Turn Speaker Utterance), each line is an Intonation Unit, space is used for tokenization, 'punctuation' are IU boundaries and should go to the end of each IU}

\item{record}{whether you want to get the step of transformation (slow process)}

\item{m}{similarity matrix to customize substitution cost}

\item{boundaries}{a list of boundary symbols that will exist in the data}

\item{noboundary}{assign a symbol for no boundary}
}
\value{
similarity score
}
\description{
Similarity score calculation
}
\examples{
m=matrix(data =
c(1,0,0,0,0,0,0,
0,1,0,0,0,0,0,
0,0,1,0,0,0,0,
0,0,0,1,0,0,0,
0,0,0,0,1,0,0,
0,0,0,0,0,1,0,
0,0,0,0,0,0,1), nrow=7)
}
